add check role Delete Catalog

+++ src/app/controllers/user.controller.js
@@ -20,6 +20,12 @@ module.exports.getAll = async (req, res) => {
 // Delete user
 module.exports.deleteUser = async (req, res, next) => {
   try {
+    const id = req.params.id;
+    if (id == "6377d09cfb12e6eafc8d1a11") {
+      return res
+        .status(501)
+        .json({ code: "501", message: "không thể xóa role admin" });
+    }
     await usersModel.findByIdAndDelete(req.params.id);
     return res
       .status(200)
@@ -233,17 +239,19 @@ exports.forgotpassword = async (req, res, next) => {
             to: user[0].email,
             subject: `Kính gửi Ông/Bà ${user[0].fullName}.`,
             text: `Xác nhận`,
-            html: `<h4>SmartSign - Khôi phục mật khẩu</h4>
-                    <body style="background:#e1e5e8" style="margin-top:0 ;margin-bottom:0 ;margin-right:0 ;margin-left:0 ;padding-top:0px;padding-bottom:0px;padding-right:0px;padding-left:0px;background-color:#e1e5e8;">
+            html: `<body  style="margin-top:0 ;margin-bottom:0 ;margin-right:0 ;margin-left:0 ;padding-top:0px;padding-bottom:0px;padding-right:0px;padding-left:0px">
             
-                    <center style="width:100%;table-layout:fixed;-webkit-text-size-adjust:100%;-ms-text-size-adjust:100%;background-color:#e1e5e8;">
-                    <div style="max-width:600px;margin-top:0;margin-bottom:0;margin-right:auto;margin-left:auto;">
+                    <center style="width:100%;table-layout:fixed;-webkit-text-size-adjust:100%;-ms-text-size-adjust:100%;margin-top:30px">
+                    <div style="max-width:580px;margin-top:0;margin-bottom:0;margin-right:auto;margin-left:auto;">
                         <table align="center" cellpadding="0" style="border-spacing:0;font-family:'Muli',Arial,sans-serif;color:#333333;Margin:0 auto;width:100%;max-width:600px;">
                         <tbody>
                             <tr>
-                            <td align="center" class="vervelogoplaceholder" height="143" style="padding-top:0;padding-bottom:0;padding-right:0;padding-left:0;height:143px;vertical-align:middle;" valign="middle"><span class="sg-image" data-imagelibrary="%7B%22width%22%3A%22160%22%2C%22height%22%3A34%2C%22alt_text%22%3A%22Verve%20Wine%22%2C%22alignment%22%3A%22%22%2C%22border%22%3A0%2C%22src%22%3A%22https%3A//marketing-image-production.s3.amazonaws.com/uploads/79d8f4f889362f0c7effb2c26e08814bb12f5eb31c053021ada3463c7b35de6fb261440fc89fa804edbd11242076a81c8f0a9daa443273da5cb09c1a4739499f.png%22%2C%22link%22%3A%22%23%22%2C%22classes%22%3A%7B%22sg-image%22%3A1%7D%7D"><a href="#" target="_blank"><img alt="Verve Wine" height="34" src="https://cf.shopee.vn/file/a56d5642e03ffd3442a0448995416f05_tn" style="border-width: 0px; width: 160px; height: 34px;" width="160"></a></span></td>
+                            <td align="center" class="vervelogoplaceholder" height="143" style="padding-top:0;padding-bottom:40px;padding-right:0;padding-left:0;height:143px;vertical-align:middle;" valign="middle"><span class="sg-image" data-imagelibrary="%7B%22width%22%3A%22160%22%2C%22height%22%3A34%2C%22alt_text%22%3A%22Verve%20Wine%22%2C%22alignment%22%3A%22%22%2C%22border%22%3A0%2C%22src%22%3A%22https%3A//marketing-image-production.s3.amazonaws.com/uploads/79d8f4f889362f0c7effb2c26e08814bb12f5eb31c053021ada3463c7b35de6fb261440fc89fa804edbd11242076a81c8f0a9daa443273da5cb09c1a4739499f.png%22%2C%22link%22%3A%22%23%22%2C%22classes%22%3A%7B%22sg-image%22%3A1%7D%7D"><a href="#" target="_blank"><img alt="Herbal oil" height="34" src="https://cf.shopee.vn/file/a56d5642e03ffd3442a0448995416f05_tn" style="border-width: 0px; width: 300px; height: 180px;" ></a></span></td>
                             </tr>
                             <!-- Start of Email Body-->
+                            <tr>
+                                <td style="text-align: center;margin-top:14px"><a href="" style="font-size: 24px;font-weight:700;letter-spacing:0.5px;text-decoration: none;color:#000000;">KHÔI PHỤC MÂT KHẨU</a></td>
+                            </tr>
                             <tr>
                             <td class="one-column" style="padding-top:0;padding-bottom:0;padding-right:0;padding-left:0;background-color:#ffffff;">
                                 <!--[if gte mso 9]>
@@ -252,21 +260,28 @@ exports.forgotpassword = async (req, res, next) => {
                                     <![endif]-->
                                 <table style="border-spacing:0;" width="100%">
                                 <tbody>
+                                   <tr>
+                                    <td><p>Gửi: ${user[0].fullName}</p></td>
+                                   </tr>
                                     <tr>
-                                    <td align="center" class="inner" style="padding-top:15px;padding-bottom:15px;padding-right:30px;padding-left:30px;" valign="middle"><span class="sg-image" data-imagelibrary="%7B%22width%22%3A%22255%22%2C%22height%22%3A93%2C%22alt_text%22%3A%22Forgot%20Password%22%2C%22alignment%22%3A%22%22%2C%22border%22%3A0%2C%22src%22%3A%22https%3A//marketing-image-production.s3.amazonaws.com/uploads/35c763626fdef42b2197c1ef7f6a199115df7ff779f7c2d839bd5c6a8c2a6375e92a28a01737e4d72f42defcac337682878bf6b71a5403d2ff9dd39d431201db.png%22%2C%22classes%22%3A%7B%22sg-image%22%3A1%7D%7D"><img alt="Forgot Password" class="banner" height="93" src="https://marketing-image-production.s3.amazonaws.com/uploads/35c763626fdef42b2197c1ef7f6a199115df7ff779f7c2d839bd5c6a8c2a6375e92a28a01737e4d72f42defcac337682878bf6b71a5403d2ff9dd39d431201db.png" style="border-width: 0px; margin-top: 30px; width: 255px; height: 93px;" width="255"></span></td>
+                                    <td class="inner contents center" style="">
+                                    <center>
+                                     <!--[if (gte mso 9)|(IE)]><![endif]-->
+                                            <p style="font-size: 14px;line-height: 1.4;text-align:left" >Chúng tôi nhân được yêu cầu đăt lại mât khẩu của bạn. Vui long nhấp vào liên kết bên dưới để thiết lập lại mật khẩu của ban!!!</p>
+                                        <div style="text-align:center;">
+<a href="" target="_blank" style="text-decoration: none;"><button alt="Reset your Password" height="54"  style="text-align: center;border:none;outline: none;cursor: pointer;background-color: #2E7D32 ;border-radius: 10px;color: #ffffff;font-weight: 700;margin-top: 12px;margin-bottom: 20px; width: 300px; height: 42px" width="260">Khôi Phục Mật Khẩu</button></a>
+                                        </div>
+                                            
+                                        <!--[if (gte mso 9)|(IE)]><br>&nbsp;<![endif]-->
+                                    </center>
+                                       
+                                    </td>
                                     </tr>
                                     <tr>
-                                    <td class="inner contents center" style="padding-top:15px;padding-bottom:15px;padding-right:30px;padding-left:30px;text-align:left;">
-                                        <center>
-                                        <p class="h1 center" style="Margin:0;text-align:center;font-family:'flama-condensed','Arial Narrow',Arial;font-weight:100;font-size:30px;Margin-bottom:26px;">Quên mật khẩu?</p>
-                                        <!--[if (gte mso 9)|(IE)]><![endif]-->
-                
-                                        <p class="description center" style="font-family:'Muli','Arial Narrow',Arial;Margin:0;text-align:center;max-width:320px;color:#a1a8ad;line-height:24px;font-size:15px;Margin-bottom:10px;margin-left: auto; margin-right: auto;"><span style="color: rgb(161, 168, 173); font-family: Muli, Arial; font-size: 15px; text-align: center; background-color: rgb(255, 255, 255);">Đừng lo lắng! Hãy truy cập vào đây để khôi phục mật khẩu</span></p>
-                                        <!--[if (gte mso 9)|(IE)]><br>&nbsp;<![endif]--><span class="sg-image" data-imagelibrary="%7B%22width%22%3A%22260%22%2C%22height%22%3A54%2C%22alt_text%22%3A%22Reset%20your%20Password%22%2C%22alignment%22%3A%22%22%2C%22border%22%3A0%2C%22src%22%3A%22https%3A//marketing-image-production.s3.amazonaws.com/uploads/c1e9ad698cfb27be42ce2421c7d56cb405ef63eaa78c1db77cd79e02742dd1f35a277fc3e0dcad676976e72f02942b7c1709d933a77eacb048c92be49b0ec6f3.png%22%2C%22link%22%3A%22%23%22%2C%22classes%22%3A%7B%22sg-image%22%3A1%7D%7D">
-                                        <a href="" target="_blank"><button alt="Reset your Password" height="54"  style="border-width: 0px;cursor: pointer;background-color: #000 ;color: #eee;margin-top: 30px; margin-bottom: 50px; width: 260px; height: 54px;" width="260">Khôi phục mật khẩu</button></a></span>
-                                         </span>
-                                        <!--[if (gte mso 9)|(IE)]><br>&nbsp;<![endif]--></center>
-                                    </td>
+                                        <td>
+                                            <p style="font-size: 14px;line-height: 1.4;">Nếu ban cần hỗ trơ thêm hoăc ban không thưc hiên thay đổi này, vui lòng liên hê: <span style="color:#2E7D32;text-decoration:underline;padding-bottom:2px;font-weight: 600;">minhphuc201.dev@gmail.com</span></p>
+                                            <span style="font-weight: 700;font-size:18px" >Tâm An Nhiên Store</span>
+                                        </td>
                                     </tr>
                                 </tbody>
                                 </table>
@@ -283,8 +298,7 @@ exports.forgotpassword = async (req, res, next) => {
                     </div>
                     </center>
             
-                
-                </body>
+</body>
                     `,
           };
 
@@ -304,4 +318,27 @@ exports.forgotpassword = async (req, res, next) => {
   } catch (err) {
     console.log(err);
   }
+};
+
+// Reset Password
+module.exports.resetPassword = async (req, res, next) => {
+  const token = req.body.token;
+  const password = req.body.password;
+  jwt.verify(token, process.env.ACCESS_TOKEN_SECRET, async (err, decode) => {
+    if (err) {
+      return res
+        .status(401)
+        .json({ code: 401, message: "Qúa thời hạn xác thực" });
+    } else {
+      const add = await usersService.NewPass(decode.userId, password);
+      if (add) {
+        return res
+          .status(200)
+          .json({ code: 200, message: "Xác thực thành công" });
+      }
+      return res
+        .status(401)
+        .json({ code: 401, message: "đổi mật khẩu thất bại" });
+    }
+  });
 };
